requirements.txt
streamlit
pandas
matplotlib
requests
ta
websocket-client
python-dotenv